Address CodeRabbit feedback for issue 2092
