+++ packages/core/src/filters/EmojiFilter.basic.test.ts
@@ -13,150 +13,6 @@ import { describe, it, expect } from 'vitest';
 import type { FilterConfiguration } from './EmojiFilter';
 import { EmojiFilter } from './EmojiFilter';
 
-// Remove mock to test actual implementation
-/*
-vi.mock('./EmojiFilter', async () => {
-  const actual = await vi.importActual('./EmojiFilter');
-
-  class MockEmojiFilter {
-    private config: FilterConfiguration;
-    private buffer: string = '';
-
-    constructor(config: FilterConfiguration) {
-      this.config = config;
-    }
-
-    filterText(text: string): any {
-      if (this.config.mode === 'allowed') {
-        return { filtered: text, emojiDetected: false, blocked: false };
-      }
-
-      const hasEmojis = /[\u{1F600}-\u{1F64F}]|[\u{1F300}-\u{1F5FF}]|[\u{1F680}-\u{1F6FF}]|[\u{1F1E0}-\u{1F1FF}]|[\u{2600}-\u{26FF}]|[\u{2700}-\u{27BF}]|✅|⚠️|🎉|🔥|💯|🚫|🎯|🤔|💭|🚀|⏳/u.test(text);
-
-      if (!hasEmojis) {
-        return { filtered: text, emojiDetected: false, blocked: false };
-      }
-
-      if (this.config.mode === 'error') {
-        return { filtered: null, emojiDetected: true, blocked: true, error: 'Emojis detected in content' };
-      }
-
-      // Warn mode - convert emojis
-      let filtered = text
-        .replace(/✅/g, '[OK]')
-        .replace(/⚠️/g, 'WARNING:')
-        .replace(/🎉|🔥|💯|🚫|🎯|🤔|💭|🚀|⏳/g, '');
-
-      return {
-        filtered,
-        emojiDetected: true,
-        blocked: false,
-        systemFeedback: 'Emojis were detected and removed. Please avoid using emojis.'
-      };
-    }
-
-    filterStreamChunk(chunk: string): any {
-      this.buffer += chunk;
-
-      // Simple implementation - only process complete words
-      if (chunk.includes(' ') || chunk === '') {
-        const result = this.filterText(this.buffer);
-        if (result.filtered !== null) {
-          this.buffer = '';
-          return result;
-        }
-      }
-
-      return { filtered: '', emojiDetected: false, blocked: false };
-    }
-
-    filterToolArgs(args: any): any {
-      const serialized = JSON.stringify(args);
-      const textResult = this.filterText(serialized);
-
-      if (textResult.blocked) {
-        return {
-          filtered: null,
-          emojiDetected: true,
-          blocked: true,
-          error: 'Cannot execute tool with emojis in parameters'
-        };
-      }
-
-      if (textResult.emojiDetected) {
-        const filteredArgs = this.deepFilterObject(args);
-        return {
-          filtered: filteredArgs,
-          emojiDetected: true,
-          blocked: false,
-          systemFeedback: 'Emojis were detected and removed from your tool call. Please avoid using emojis in tool parameters.'
-        };
-      }
-
-      return { filtered: args, emojiDetected: false, blocked: false };
-    }
-
-    filterFileContent(content: string, toolName: string): any {
-      const result = this.filterText(content);
-
-      if (result.blocked) {
-        return {
-          filtered: null,
-          emojiDetected: true,
-          blocked: true,
-          error: 'Cannot write emojis to code files'
-        };
-      }
-
-      if (result.emojiDetected) {
-        return {
-          ...result,
-          systemFeedback: `Emojis were removed from ${toolName} content. Please avoid using emojis in code.`
-        };
-      }
-
-      return result;
-    }
-
-    flushBuffer(): string {
-      if (this.config.mode === 'error' && /[\u{1F600}-\u{1F64F}]|[\u{1F300}-\u{1F5FF}]|[\u{1F680}-\u{1F6FF}]|[\u{1F1E0}-\u{1F1FF}]|[\u{2600}-\u{26FF}]|[\u{2700}-\u{27BF}]|✅|⚠️|🎉|🔥|💯|🚫|🎯|🤔|💭|🚀|⏳/u.test(this.buffer)) {
-        this.buffer = '';
-        return '';
-      }
-
-      const result = this.filterText(this.buffer);
-      this.buffer = '';
-      return result.filtered || '';
-    }
-
-    private deepFilterObject(obj: any): any {
-      if (typeof obj === 'string') {
-        return this.filterText(obj).filtered;
-      }
-
-      if (Array.isArray(obj)) {
-        return obj.map(item => this.deepFilterObject(item));
-      }
-
-      if (obj && typeof obj === 'object') {
-        const filtered: any = {};
-        for (const [key, value] of Object.entries(obj)) {
-          filtered[key] = this.deepFilterObject(value);
-        }
-        return filtered;
-      }
-
-      return obj;
-    }
-  }
-
-  return {
-    ...actual,
-    EmojiFilter: MockEmojiFilter
-  };
-});
-*/
-
 describe('EmojiFilter', () => {
   describe('constructor', () => {
     it('should create filter with allowed mode configuration', () => {
